+++ nn/src/loss/loss.c
@@ -54,7 +54,7 @@ double mean_squared_error(const Matrix* y_hat, const Matrix* y) {
   double loss = 0.0;
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     double diff = y_hat->matrix_data[i] - y->matrix_data[i];
     loss += pow(diff, 2);
   }
@@ -69,7 +69,7 @@ double categorical_cross_entropy(const Matrix* y_hat, const Matrix* y) {
   double loss = 0.0;
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     loss -= y->matrix_data[i] * log(y_hat->matrix_data[i] + EPSILON);
   }
 
@@ -83,7 +83,7 @@ double mean_absolute_error(const Matrix* y_hat, const Matrix* y) {
   double loss = 0.0;
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     loss += fabs(y_hat->matrix_data[i] - y->matrix_data[i]);
   }
 
@@ -97,7 +97,7 @@ double binary_cross_entropy(const Matrix* y_hat, const Matrix* y) {
   double loss = 0.0;
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     loss -= y->matrix_data[i] * log(y_hat->matrix_data[i] + EPSILON) +
             (1 - y->matrix_data[i]) * log(1 - y_hat->matrix_data[i] + EPSILON);
   }
@@ -112,7 +112,7 @@ Matrix* mean_squared_error_gradient(const Matrix* y_hat, const Matrix* y) {
   Matrix* gradient = create_matrix(y_hat->rows, y_hat->cols);
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     gradient->matrix_data[i] =
         2.0 * (y_hat->matrix_data[i] - y->matrix_data[i]);
   }
@@ -129,7 +129,7 @@ Matrix* categorical_cross_entropy_gradient(const Matrix* y_hat,
   Matrix* gradient = create_matrix(y_hat->rows, y_hat->cols);
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     gradient->matrix_data[i] =
         -y->matrix_data[i] / (y_hat->matrix_data[i] + EPSILON);
   }
@@ -144,7 +144,7 @@ Matrix* mean_absolute_error_gradient(const Matrix* y_hat, const Matrix* y) {
   Matrix* gradient = create_matrix(y_hat->rows, y_hat->cols);
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     if (y_hat->matrix_data[i] > y->matrix_data[i]) {
       gradient->matrix_data[i] = 1.0;
     } else if (y_hat->matrix_data[i] < y->matrix_data[i]) {
@@ -165,7 +165,7 @@ Matrix* binary_cross_entropy_gradient(const Matrix* y_hat, const Matrix* y) {
   Matrix* gradient = create_matrix(y_hat->rows, y_hat->cols);
   int total_elements = y_hat->rows * y_hat->cols;
 
-  for (int i = 0; i < total_elements; i++) {
+  for (size_t i = 0; i < total_elements; i++) {
     gradient->matrix_data[i] =
         (y_hat->matrix_data[i] - y->matrix_data[i]) /
         (y_hat->matrix_data[i] * (1 - y_hat->matrix_data[i]) + EPSILON);
